pyproject.toml
[tool.poetry]
name = "ao3-rss"
version = "0.1.0"
description = "Follow AO3 works without making an account."
authors = ["Kyle Anthony Williams <kyle.anthony.williams2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
rfeed = "^1.1.1"
ao3-api = "^2.2.1"
requests = "^2.27.1"


[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
